Fleshing out SVG API
Reworked attribute handling code to accommodate custom converters.

That capability is used to implement short-hands for SVG transforms —
and in a later change polyline-s & path-s

Also had to fix a bug with leaf handling support when the leaf element
is at the end of contents section.  That caused a bug with improperly
nesting SVG g-roups.